Rename spinner package
Prior to this commit `CharMap` was used to describe the animations that
could be used with a spinner. The UX for this wasn't great because it
wasn't really clear what a `CharMap` was.

This commit renames the spinner package to Animations. Types and methods
have also been renamed to match.