Fix incorrect line counts
Prior to this commit, if `AddSpinner()` was called after `Start()` and
the number of spinners exceeded the lines left on the screen rendering
would break.
This caused frames to be written to a new line for each frame.

This commit fixes that by ensuring that the screen is buffered inside
the processing loop. The result is that we can recalculate the number of
lines needed for each render cycle.