Fix line updates
Prior to this commit, lines were not properly cleared. This meant that
updating the message of a spinner could result in text being incorrectly
displayed. This would have been more noticeable when the updated message
was shorted than the original message.

This commit fixes that by adding a new ClearLine method to the tput
package and using it to ensure that each line is cleared before it is
rendered.