Enable realtime message updates
Prior to this commit spinner messages updates were rendered along with
the spinner frames.

This commit adds support for realtime message updates by using a non
blocking channel to notify the manager that there has been a message
update.

When this event is true, only the message is rendered. The
spinner frame will remain at it's previous position.