Add UpdateMessagef() method
Prior to this commit it was not possible to pass a formatted string to
update the message of a spinner. The consumer would have to assign the
result of fmt.Sprintf() to a variable and then pass it as a parameter to
UpdateMessage().

This commit adds the UpdateMessagef() method so that consumers can
update spinner messages with a formatted string.