Add default spinner speeds
Prior to this commit spinners would use the default speed of 250
milliseconds set in the managers constructor method.

This commit introduces a new concept of animation properties that allow
each spinner to have it's own default speed. The speeds used come from
https://wiki.tcl-lang.org/page/Text+Spinner.

It is still possible to globally override the speed via the managers
WithFrameDuration option.