Enhance Complete and Error methods
Prior to this change a user of this package would need to call
UpdateMessage if they wanted to update the spinner with a message
after either Complete or Error had been called.

This change introduces four new functions that will allow users to
control messages when marking a spinner as Complete or in an Error
state.

```
ErrorWithMessage(..)
ErrorWithMessagef(..)
CompleteWithMessage(..)
CompleteWithMessagef(..)
```